+++ game/client/c_baseentity.h
@@ -1513,9 +1513,9 @@ private:
 	void MoveToAimEnt( );
 	
 	// Sets/Gets the next think based on context index
-	void SetNextThink( int nContextIndex, float thinkTime );
-	void SetLastThink( int nContextIndex, float thinkTime );
-	float GetNextThink( int nContextIndex ) const;
+	void SetNextThink( int nContextIndex, double thinkTime );
+	void SetLastThink( int nContextIndex, double thinkTime );
+	double GetNextThink( int nContextIndex ) const;
 	int	GetNextThinkTick( int nContextIndex ) const;
 
 	// Object velocity
