+++ game/client/physics.h
@@ -22,6 +22,6 @@ struct solid_t;
 extern bool PhysicsDLLInit( CreateInterfaceFn physicsFactory );
 extern void PhysicsReset();
 extern void PhysicsSimulate();
-extern float PhysGetSyncCreateTime();
+extern double PhysGetSyncCreateTime();
 
 #endif // PHYSICS_H
